# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /subtract:
    post:
      tags:
        - math-solution-controller
      operationId: subtractOperators
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /multiply:
    post:
      tags:
        - math-solution-controller
      operationId: multiplyOperators
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /divide:
    post:
      tags:
        - math-solution-controller
      operationId: divideOperators
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /add:
    post:
      tags:
        - math-solution-controller
      operationId: addOperators
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /randomMonth:
    get:
      tags:
        - month-controller
      operationId: randomMonth
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /month/{monthNumber}:
    get:
      tags:
        - month-controller
      operationId: convertMonth
      parameters:
        - name: monthNumber
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    MathSolution:
      required:
        - operand1
        - operand2
      type: object
      properties:
        operand1:
          minimum: -999
          type: number
          format: double
        operand2:
          minimum: -999
          type: number
          format: double
        operation:
          type: string
        answer:
          type: number
          format: double
    Month:
      type: object
      properties:
        number:
          maximum: 12
          minimum: 1
          type: integer
          format: int32
        name:
          type: string